requirements.txt
python-dotenv
openai
xmltodict
pandas